refactor(chat): split composer attachments
